fix: enabling disabled tests rename all pkgs to lib
