+++ packages/binding-abstract/lib/index.js
