+++ packages/binding-mock/lib/index.js
