+++ packages/binding-mock/lib/legacy.js
