+++ packages/parser-byte-length/lib/index.js
