+++ packages/parser-cctalk/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@serialport/parser-cctalk",
   "version": "8.0.1",
-  "main": "cctalk.js",
+  "main": "lib",
   "engines": {
     "node": ">=8.6.0"
   },
