+++ packages/parser-delimiter/lib/index.js
