+++ packages/parser-inter-byte-timeout/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@serialport/parser-inter-byte-timeout",
   "version": "8.0.1",
-  "main": "inter-byte-timeout.js",
+  "main": "lib",
   "engines": {
     "node": ">=8.6.0"
   },
