+++ packages/parser-readline/lib/index.js
