+++ packages/parser-ready/lib/index.js
