+++ packages/parser-ready/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@serialport/parser-ready",
-  "main": "ready.js",
+  "main": "lib",
   "version": "8.0.1",
   "engines": {
     "node": ">=8.6.0"
