+++ packages/parser-regex/lib/index.js
