+++ packages/parser-regex/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@serialport/parser-regex",
-  "main": "regex.js",
+  "main": "lib",
   "version": "8.0.1",
   "engines": {
     "node": ">=8.6.0"
