+++ packages/parser-slip-encoder/lib/index.js
