+++ packages/parser-slip-encoder/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@serialport/parser-slip-encoder",
-  "main": "slip-encoder.js",
+  "main": "lib",
   "version": "8.0.1",
   "engines": {
     "node": ">=8.6.0"
