+++ packages/stream/lib/index.js
