+++ packages/terminal/lib/index.js
