+++ test/initializers/index.js
@@ -1,3 +1,6 @@
 const fs = require('fs')
 const files = fs.readdirSync(__dirname).filter(file => file !== 'index.js' && file.match(/\.js$/))
-files.forEach(file => require(`./${file}`))
+files.forEach(file => {
+  console.log(`Loading ${file}`)
+  require(`./${file}`)
+})
