+++ test/mocha.opts
@@ -1,5 +1,3 @@
 --bail
 --require ./test/initializers
-packages/*/test/**/*.js
 packages/**/*.test.js
-
